requirements.txt
seaborn
pandas
matplotlib
pyyaml
inquirerpy
